test_requirements.txt
django-nose
flake8
# coverage
# django-coverage
